rewrite first page
