+++ css/style.css
@@ -13,7 +13,7 @@ body{ overflow: hidden; }
 }
 
 #map-body{
-	margin-top: 5%;
+
 	height: 85%;
 	width: 100%;
 }
