+++ models/Icons.js
@@ -1,0 +1,23 @@
+export const startIcon = L.icon.mapkey(
+  {
+    icon: 'avatar',
+    color: '#1B8717',
+    background: 'white',
+    size: 35,
+  });
+
+export const endIcon = L.icon.mapkey(
+  {
+    icon: 'golf',
+    color: 'red',
+    background: 'white',
+    size: 35,
+  });
+
+export const busIcon = L.icon.mapkey(
+  {
+    icon: 'bus',
+    color: 'white',
+    background: '#221961',
+    size: 30,
+  });
