+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/pub-t/plain-destination#readme",
   "devDependencies": {
+    "babel": "^6.5.2",
     "concurrently": "^2.1.0",
     "eslint": "^2.12.0",
     "eslint-config-airbnb": "^9.0.1",
